update StoryblokProvider and getAllStories
